Move reminder-related commands and tasks to cogs/reminders.py
